Feature: enable/disable permalinks for RSS The option to see the shortlinks or permalinks has been added to the configuration panel. It is a simple checkbox This option is disabled by default (meaning that shortlinks are the default) Updated writeConfig() to save this option Also fixed a slight typo in config.html.
Removed useless CSS & fixed a comment

Enabled permalinks for the ATOM feed and fixed the isPermaLink attribute for the <guid> tag

Reverted to default behavior and clarified its meaning
EnableRssPermalinks is an oddly behaving option: when enabled, it shows a
permalink in the description and a full link in the element title, and
swaps it around when disabled. This clarifies the option for end-users
Also, moved enable_rss_permalinks to $GLOBALS['config'] because it is a
config option.

fix indent